+++ NEWS.md
@@ -55,8 +55,16 @@
    [#2574](https://github.com/JuliaData/DataFrames.jl/pull/2574),
    [#2664](https://github.com/JuliaData/DataFrames.jl/pull/2664))
 
-# DataFrames v0.22.6 Release Notes
+# DataFrames v0.22.6 Release notes
 
+* `convert` methods from `AbstractDataFrame`, `DataFrameRow` and `GroupKey`
+  to `Array`, `Matrix`, `Vector` and `Tuple`, as well as from `AbstractDict` to
+  `DataFrame`, are now deprecated: use corresponding
+  constructors instead. The only conversions that are
+  retained are `convert(::Type{NamedTuple}, dfr::DataFrameRow)`,
+  `convert(::Type{NamedTuple}, key::GroupKey)`, and
+  `convert(::Type{DataFrame}, sdf::SubDataFrame)`; the deprecated methods will be
+  removed in 1.0 release
 * as a bug fix `eltype` of vector returned by `eachrow` is now `DataFrameRow`
   ([#2662](https://github.com/JuliaData/DataFrames.jl/pull/2662))
 * applying `map` to `GroupedDataFrame` is now deprecated. It will